@@ -5,20 +5,23 @@ version = "0.10.5"
 enable_black = true
 enable_flake8 = true
 enable_isort = true
-enable_mypy = false    # TODO: eliminate errors and enable at CI
-mypy_preset = "entry"  # TODO: "strict"
+enable_mypy = true
+mypy_preset = "entry" # TODO: "strict"
+mypy_plugins = [
+  { function = "numpy.typing.mypy_plugin" },
+  { function = "pydantic.mypy" },
+]
 line_length = 88
 py_version = "py311"
 isort_known_first_party = ["voicevox_engine"]
 isort_known_third_party = ["numpy"]
 [[tool.pysen.lint.mypy_targets]]
-  paths = [".", "voicevox_engine/"]
+paths = ["."]
 
 [tool.black] # automatically generated by pysen
 # pysen ignores and overwrites any modifications
 line-length = 88
-target-version = ["py310", "py311"]
-
+target-version = ["py311"]
 
 [tool.isort] # automatically generated by pysen
 # pysen ignores and overwrites any modifications
@@ -45,23 +48,18 @@ numpy = "^1.20.0"
 fastapi = "^0.103.2"
 python-multipart = "^0.0.5"
 uvicorn = "^0.15.0"
-aiofiles = "^0.7.0"
 soundfile = "^0.12.1"
 pyyaml = "^6.0"
 pyworld = "^0.3.0"
-requests = "^2.28.1"
 jinja2 = "^3.1.2"
-pyopenjtalk = {git = "https://github.com/VOICEVOX/pyopenjtalk", rev = "b35fc89fe42948a28e33aed886ea145a51113f88"}
+pyopenjtalk = { git = "https://github.com/VOICEVOX/pyopenjtalk", rev = "b35fc89fe42948a28e33aed886ea145a51113f88" }
 semver = "^3.0.0"
 platformdirs = "^3.10.0"
 soxr = "^0.3.6"
 
 [tool.poetry.group.dev.dependencies]
-cython = "^0.29.34,>=0.29.33" # NOTE: for Python 3.11
-pyinstaller = "^5.13"
+pyinstaller = "^6.2.0"
 pre-commit = "^2.16.0"
-atomicwrites = "^1.4.0"
-colorama = "^0.4.4"
 poetry = "^1.3.1"
 
 [tool.poetry.group.test.dependencies]
@@ -71,10 +69,12 @@ flake8-bugbear = "^23.1.0"
 flake8 = "^6.0.0"
 isort = "^5.12.0"
 mypy = "^1.6.0"
-pytest = "^6.2.5"
+pytest = "^7.4.3"
 coveralls = "^3.2.0"
 poetry = "^1.3.1"
-httpx = "^0.25.0" # NOTE: required by fastapi.testclient.TestClient
+httpx = "^0.25.0"          # NOTE: required by fastapi.testclient.TestClient
+syrupy = "^4.6.0"
+types-pyyaml = "^6.0"
 
 [tool.poetry.group.license.dependencies]
 pip-licenses = "^4.2.0"
