@@ -2,7 +2,6 @@ import json
 from copy import deepcopy
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import Dict
 from unittest import TestCase
 
 from fastapi import HTTPException
@@ -11,8 +10,8 @@ from pyopenjtalk import g2p, unset_user_dict
 from voicevox_engine.model import UserDictWord, WordTypes
 from voicevox_engine.part_of_speech_data import MAX_PRIORITY, part_of_speech_data
 from voicevox_engine.user_dict import (
+    _create_word,
     apply_word,
-    create_word,
     delete_word,
     import_user_dict,
     read_dict,
@@ -61,7 +60,7 @@ import_word = UserDictWord(
 )
 
 
-def get_new_word(user_dict: Dict[str, UserDictWord]):
+def get_new_word(user_dict: dict[str, UserDictWord]) -> UserDictWord:
     assert len(user_dict) == 2 or (
         len(user_dict) == 1 and "aab7dda2-0d97-43c8-8cb7-3f440dab9b4e" not in user_dict
     )
@@ -90,7 +89,7 @@ class TestUserDict(TestCase):
     def test_create_word(self):
         # 将来的に品詞などが追加された時にテストを増やす
         self.assertEqual(
-            create_word(surface="test", pronunciation="テスト", accent_type=1),
+            _create_word(surface="test", pronunciation="テスト", accent_type=1),
             UserDictWord(
                 surface="ｔｅｓｔ",
                 priority=5,
@@ -219,7 +218,7 @@ class TestUserDict(TestCase):
         for pos in part_of_speech_data:
             for i in range(MAX_PRIORITY + 1):
                 self.assertEqual(
-                    create_word(
+                    _create_word(
                         surface="test",
                         pronunciation="テスト",
                         accent_type=1,
