@@ -1,14 +1,16 @@
 import json
 import sys
 from pathlib import Path
-from typing import Dict, List, Optional
+from typing import List, Optional
 
-from ..utility import engine_root, get_save_dir
 from .core_wrapper import CoreWrapper, load_runtime_lib
-from .synthesis_engine import SynthesisEngine, SynthesisEngineBase
+from .tts_pipeline.tts_engine import CoreAdapter
+from .utility import engine_root, get_save_dir
+
+MOCK_VER = "0.0.0"
 
 
-def make_synthesis_engines(
+def initialize_cores(
     use_gpu: bool,
     voicelib_dirs: Optional[List[Path]] = None,
     voicevox_dir: Optional[Path] = None,
@@ -16,9 +18,9 @@ def make_synthesis_engines(
     cpu_num_threads: Optional[int] = None,
     enable_mock: bool = True,
     load_all_models: bool = False,
-) -> Dict[str, SynthesisEngineBase]:
+) -> dict[str, CoreAdapter]:
     """
-    音声ライブラリをロードして、音声合成エンジンを生成
+    音声ライブラリをロードしてコアを生成
 
     Parameters
     ----------
@@ -47,6 +49,8 @@ def make_synthesis_engines(
         )
         cpu_num_threads = 0
 
+    # ディレクトリを設定する
+    # 引数による指定を反映する
     if voicevox_dir is not None:
         if voicelib_dirs is not None:
             voicelib_dirs.append(voicevox_dir)
@@ -63,40 +67,54 @@ def make_synthesis_engines(
         if runtime_dirs is None:
             runtime_dirs = [root_dir]
 
+    # `~`をホームディレクトリのパスに置き換える
     voicelib_dirs = [p.expanduser() for p in voicelib_dirs]
     runtime_dirs = [p.expanduser() for p in runtime_dirs]
 
+    # ランタイムをロードする
     load_runtime_lib(runtime_dirs)
 
-    synthesis_engines = {}
+    # コアをロードし `cores` へ登録する
+    cores: dict[str, CoreAdapter] = {}
 
     if not enable_mock:
 
-        def load_core_library(core_dir: Path, suppress_error: bool = False):
+        def load_core_library(core_dir: Path, suppress_error: bool = False) -> None:
             """
-            指定されたディレクトリにあるコアを読み込む。
-            ユーザーディレクトリの場合は存在しないこともあるので、エラーを抑制すると良い。
+            指定されたコアをロードし `cores` へ登録する。
+            Parameters
+            ----------
+            core_dir : Path
+                直下にコア（共有ライブラリ）が存在するディレクトリ、あるいはその候補
+            suppress_error: bool
+                エラーを抑制する。`core_dir` がコア候補であることを想定。
             """
+            # 指定されたコアをロードし登録する
             try:
+                # コアをロードする
                 core = CoreWrapper(use_gpu, core_dir, cpu_num_threads, load_all_models)
+                # コアを登録する
                 metas = json.loads(core.metas())
                 core_version = metas[0]["version"]
                 print(f"Info: Loading core {core_version}.")
-                if core_version in synthesis_engines:
+                if core_version in cores:
                     print(
                         "Warning: Core loading is skipped because of version duplication.",
                         file=sys.stderr,
                     )
                 else:
-                    synthesis_engines[core_version] = SynthesisEngine(core=core)
+                    cores[core_version] = CoreAdapter(core)
             except Exception:
+                # コアでなかった場合のエラーを抑制する
                 if not suppress_error:
                     raise
 
+        # `voicelib_dirs` 下のコアをロードし登録する
         for core_dir in voicelib_dirs:
             load_core_library(core_dir)
 
-        # ユーザーディレクトリにあるコアを読み込む
+        # ユーザーディレクトリ下のコアをロードし登録する
+        # コア候補を列挙する
         user_voicelib_dirs = []
         core_libraries_dir = get_save_dir() / "core_libraries"
         core_libraries_dir.mkdir(exist_ok=True)
@@ -105,20 +123,17 @@ def make_synthesis_engines(
             if not path.is_dir():
                 continue
             user_voicelib_dirs.append(path)
-
+        # コア候補をロードし登録する。候補がコアで無かった場合のエラーを抑制する。
         for core_dir in user_voicelib_dirs:
             load_core_library(core_dir, suppress_error=True)
 
     else:
         # モック追加
-        from ..dev.core import metas as mock_metas
-        from ..dev.core import supported_devices as mock_supported_devices
-        from ..dev.synthesis_engine import MockSynthesisEngine
+        from .dev.core import MockCoreWrapper
 
-        if "0.0.0" not in synthesis_engines:
+        if MOCK_VER not in cores:
             print("Info: Loading mock.")
-            synthesis_engines["0.0.0"] = MockSynthesisEngine(
-                speakers=mock_metas(), supported_devices=mock_supported_devices()
-            )
+            core = MockCoreWrapper()
+            cores[MOCK_VER] = CoreAdapter(core)
 
-    return synthesis_engines
+    return cores
