@@ -3,9 +3,8 @@ import json
 import os
 import shutil
 import zipfile
-from io import BytesIO
 from pathlib import Path
-from typing import Dict
+from typing import BinaryIO
 
 from fastapi import HTTPException
 from pydantic import ValidationError
@@ -23,6 +22,8 @@ INFO_FILE = "metas.json"
 
 
 class LibraryManager:
+    """音声ライブラリ (`.vvlib`) の管理"""
+
     def __init__(
         self,
         library_root_dir: Path,
@@ -42,7 +43,13 @@ class LibraryManager:
         self.engine_name = engine_name
         self.engine_uuid = engine_uuid
 
-    def downloadable_libraries(self):
+    def downloadable_libraries(self) -> list[DownloadableLibraryInfo]:
+        """
+        ダウンロード可能ライブラリの一覧を取得
+        Returns
+        -------
+        - : list[DownloadableLibraryInfo]
+        """
         # == ダウンロード情報をネットワーク上から取得する場合
         # url = "https://example.com/downloadable_libraries.json"
         # response = requests.get(url)
@@ -82,18 +89,39 @@ class LibraryManager:
                 ]
             return list(map(DownloadableLibraryInfo.parse_obj, libraries))
 
-    def installed_libraries(self) -> Dict[str, InstalledLibraryInfo]:
-        library = {}
+    def installed_libraries(self) -> dict[str, InstalledLibraryInfo]:
+        """
+        インストール済み音声ライブラリの情報を取得
+        Returns
+        -------
+        library : Dict[str, InstalledLibraryInfo]
+            インストール済みライブラリの情報
+        """
+        library: dict[str, InstalledLibraryInfo] = {}
         for library_dir in self.library_root_dir.iterdir():
             if library_dir.is_dir():
+                # ライブラリ情報の取得 from `library_root_dir / f"{library_uuid}" / "metas.json"`
                 library_uuid = os.path.basename(library_dir)
                 with open(library_dir / INFO_FILE, encoding="utf-8") as f:
-                    library[library_uuid] = json.load(f)
-                    # アンインストール出来ないライブラリを作る場合、何かしらの条件でFalseを設定する
-                    library[library_uuid]["uninstallable"] = True
+                    info = json.load(f)
+                # アンインストール出来ないライブラリを作る場合、何かしらの条件でFalseを設定する
+                library[library_uuid] = InstalledLibraryInfo(**info, uninstallable=True)
         return library
 
-    def install_library(self, library_id: str, file: BytesIO):
+    def install_library(self, library_id: str, file: BinaryIO) -> Path:
+        """
+        音声ライブラリ (`.vvlib`) のインストール
+        Parameters
+        ----------
+        library_id : str
+            インストール対象ライブラリID
+        file : BytesIO
+            ライブラリファイルBlob
+        Returns
+        -------
+        library_dir : Path
+            インストール済みライブラリの情報
+        """
         for downloadable_library in self.downloadable_libraries():
             if downloadable_library.uuid == library_id:
                 library_info = downloadable_library.dict()
@@ -102,10 +130,16 @@ class LibraryManager:
             raise HTTPException(
                 status_code=404, detail=f"指定された音声ライブラリ {library_id} が見つかりません。"
             )
+
+        # ライブラリディレクトリの生成
         library_dir = self.library_root_dir / library_id
         library_dir.mkdir(exist_ok=True)
+
+        # metas.jsonの生成
         with open(library_dir / INFO_FILE, "w", encoding="utf-8") as f:
             json.dump(library_info, f, indent=4, ensure_ascii=False)
+
+        # zipファイル形式のバリデーション
         if not zipfile.is_zipfile(file):
             raise HTTPException(
                 status_code=422, detail=f"音声ライブラリ {library_id} は不正なファイルです。"
@@ -117,7 +151,7 @@ class LibraryManager:
                     status_code=422, detail=f"音声ライブラリ {library_id} は不正なファイルです。"
                 )
 
-            # validate manifest version
+            # マニフェストファイルの存在とファイル形式をバリデーション
             vvlib_manifest = None
             try:
                 vvlib_manifest = json.loads(
@@ -134,6 +168,7 @@ class LibraryManager:
                     detail=f"指定された音声ライブラリ {library_id} のvvlib_manifest.jsonは不正です。",
                 )
 
+            # マニフェスト形式のバリデーション
             try:
                 VvlibManifest.validate(vvlib_manifest)
             except ValidationError:
@@ -142,11 +177,13 @@ class LibraryManager:
                     detail=f"指定された音声ライブラリ {library_id} のvvlib_manifest.jsonに不正なデータが含まれています。",
                 )
 
+            # ライブラリバージョンのバリデーション
             if not Version.is_valid(vvlib_manifest["version"]):
                 raise HTTPException(
                     status_code=422, detail=f"指定された音声ライブラリ {library_id} のversionが不正です。"
                 )
 
+            # マニフェストバージョンのバリデーション
             try:
                 vvlib_manifest_version = Version.parse(
                     vvlib_manifest["manifest_version"]
@@ -156,33 +193,45 @@ class LibraryManager:
                     status_code=422,
                     detail=f"指定された音声ライブラリ {library_id} のmanifest_versionが不正です。",
                 )
-
             if vvlib_manifest_version > self.supported_vvlib_version:
                 raise HTTPException(
                     status_code=422, detail=f"指定された音声ライブラリ {library_id} は未対応です。"
                 )
 
+            # ライブラリ-エンジン対応のバリデーション
             if vvlib_manifest["engine_uuid"] != self.engine_uuid:
                 raise HTTPException(
                     status_code=422,
                     detail=f"指定された音声ライブラリ {library_id} は{self.engine_name}向けではありません。",
                 )
 
+            # 展開によるインストール
             zf.extractall(library_dir)
+
         return library_dir
 
-    def uninstall_library(self, library_id: str):
+    def uninstall_library(self, library_id: str) -> None:
+        """
+        インストール済み音声ライブラリのアンインストール
+        Parameters
+        ----------
+        library_id : str
+            インストール対象ライブラリID
+        """
+        # 対象ライブラリがインストール済みであることの確認
         installed_libraries = self.installed_libraries()
         if library_id not in installed_libraries.keys():
             raise HTTPException(
                 status_code=404, detail=f"指定された音声ライブラリ {library_id} はインストールされていません。"
             )
 
-        if not installed_libraries[library_id]["uninstallable"]:
+        # アンインストール許可フラグのバリデーション
+        if not installed_libraries[library_id].uninstallable:
             raise HTTPException(
                 status_code=403, detail=f"指定された音声ライブラリ {library_id} はアンインストールできません。"
             )
 
+        # ディレクトリ削除によるアンインストール
         try:
             shutil.rmtree(self.library_root_dir / library_id)
         except Exception:
