@@ -3,7 +3,7 @@ from fastapi.testclient import TestClient
 from voicevox_engine import __version__
 
 
-def test_fetch_version_success(client: TestClient):
+def test_fetch_version_success(client: TestClient) -> None:
     response = client.get("/version")
     assert response.status_code == 200
     assert response.json() == __version__
