@@ -1,17 +1,3 @@
-from .mock import (
-    decode_forward,
-    initialize,
-    metas,
-    supported_devices,
-    yukarin_s_forward,
-    yukarin_sa_forward,
-)
+from .mock import MockCoreWrapper
 
-__all__ = [
-    "decode_forward",
-    "initialize",
-    "yukarin_s_forward",
-    "yukarin_sa_forward",
-    "metas",
-    "supported_devices",
-]
+__all__ = ["MockCoreWrapper"]
