@@ -6,7 +6,7 @@ from pydantic import BaseModel, Field
 
 class SpeakerStyle(BaseModel):
     """
-    スピーカーのスタイル情報
+    話者のスタイル情報
     """
 
     name: str = Field(title="スタイル名")
@@ -35,28 +35,28 @@ class SpeakerSupportedFeatures(BaseModel):
 
 class CoreSpeaker(BaseModel):
     """
-    コアに含まれるスピーカー情報
+    コアに含まれる話者情報
     """
 
     name: str = Field(title="名前")
-    speaker_uuid: str = Field(title="スピーカーのUUID")
-    styles: List[SpeakerStyle] = Field(title="スピーカースタイルの一覧")
-    version: str = Field("スピーカーのバージョン")
+    speaker_uuid: str = Field(title="話者のUUID")
+    styles: List[SpeakerStyle] = Field(title="スタイルの一覧")
+    version: str = Field("話者のバージョン")
 
 
 class EngineSpeaker(BaseModel):
     """
-    エンジンに含まれるスピーカー情報
+    エンジンに含まれる話者情報
     """
 
     supported_features: SpeakerSupportedFeatures = Field(
-        title="スピーカーの対応機能", default_factory=SpeakerSupportedFeatures
+        title="話者の対応機能", default_factory=SpeakerSupportedFeatures
     )
 
 
 class Speaker(CoreSpeaker, EngineSpeaker):
     """
-    スピーカー情報
+    話者情報
     """
 
     pass
