@@ -8,7 +8,7 @@ class Preset(BaseModel):
 
     id: int = Field(title="プリセットID")
     name: str = Field(title="プリセット名")
-    speaker_uuid: str = Field(title="スピーカーのUUID")
+    speaker_uuid: str = Field(title="話者のUUID")
     style_id: int = Field(title="スタイルID")
     speedScale: float = Field(title="全体の話速")
     pitchScale: float = Field(title="全体の音高")
