@@ -5,7 +5,7 @@ from .connect_base64_waves import (
 )
 from .core_version_utility import get_latest_core_version, parse_core_version
 from .mutex_utility import mutex_wrapper
-from .path_utility import delete_file, engine_root, get_save_dir
+from .path_utility import delete_file, engine_root, get_save_dir, internal_root
 
 __all__ = [
     "ConnectBase64WavesException",
@@ -13,8 +13,9 @@ __all__ = [
     "decode_base64_waves",
     "get_latest_core_version",
     "parse_core_version",
+    "mutex_wrapper",
     "delete_file",
     "engine_root",
     "get_save_dir",
-    "mutex_wrapper",
+    "internal_root",
 ]
